naming and structure changes
