+++ src/components/common/Allowance/AllowanceModal/index.ts
@@ -1,0 +1,3 @@
+import Wrapper from "./Wrapper"
+
+export default Wrapper
