+++ src/components/common/Allowance/hooks/useAllowance.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from "react"
+
+const AllowanceContext = createContext(null)
+
+const useAllowance = () => useContext(AllowanceContext)
+
+export { useAllowance, AllowanceContext }
