+++ src/components/common/Allowance/index.ts
@@ -1,5 +1,5 @@
+import AllowanceModal from "./AllowanceModal"
 import AllowanceModalBody from "./AllowanceModalBody"
 import AllowanceModalFooter from "./AllowanceModalFooter"
-import AllowanceModal from "./Modal"
 
 export { AllowanceModal, AllowanceModalBody, AllowanceModalFooter }
