--- src/components/common/AllowanceModal/Modal/index.ts
@@ -1,3 +1,0 @@
-import Modal from "./Modal"
-
-export default Modal
